requirements.txt
scrapy==1.0.3
pyquery==1.2.9
ftfy==4.0.0
jsonschema==2.5.1
python_dateutil==2.4.2
nltk==3.1
